Use proper type

+++ SkillShotMode.h
@@ -3,9 +3,9 @@
 
 class SkillShot {
   private:
-    byte activeRollover_;
-    byte bonusValue_;
-    byte scoreValue_;
+    byte          activeRollover_;
+    byte          bonusValue_;
+    unsigned long scoreValue_;
 
     byte checkRollover(byte switchHit);
     void handleNewMode();
